requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.134.0
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
